refactor: move make_donor_dataset to philanthropy.datasets with a deprecation shim
Closes #111. The gift-level generator now lives beside
generate_synthetic_donor_data; the utils path keeps a working shim that
emits exactly one DeprecationWarning (removed in 0.8.0), registered in
tests/test_deprecations.py per RELEASING.md. tests/conftest.py switches
to the new location so the suite itself stops exercising the deprecated
path. Note: point 4 of the issue (missing docstring) was already
resolved on main, so the docstring only needed its example re-pointed.
