+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased]
 
+### Deprecated
+- `philanthropy.utils.make_donor_dataset` moves to
+  [`philanthropy.datasets.make_donor_dataset`](philanthropy/datasets/) and the
+  old location emits a `DeprecationWarning`; removed in 0.8.0. The gift-level
+  generator now lives next to `generate_synthetic_donor_data`, which is the
+  canonical datasets home. Closes #111.
+
 ### Changed
 - `GratefulPatientFeaturizer`, `EncounterTransformer` and the `philanthropy`
   CLI now reject network-scheme paths (`https://`, `s3://`, `gs://`) with a
