+++ philanthropy/datasets/__init__.py
@@ -5,6 +5,10 @@ Synthetic data generators and real reference datasets for donor analytics.
 """
 
 from ._ciob import load_ciob_fundraising
-from ._generator import generate_synthetic_donor_data
+from ._generator import generate_synthetic_donor_data, make_donor_dataset
 
-__all__ = ["generate_synthetic_donor_data", "load_ciob_fundraising"]
+__all__ = [
+    "generate_synthetic_donor_data",
+    "make_donor_dataset",
+    "load_ciob_fundraising",
+]
